SocialNetwork/requirements.txt
aiosignal==1.3.1
asgiref==3.6.0
astroid==2.15.4
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
colorama==0.4.6
dill==0.3.6
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
platformdirs==3.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tzdata==2023.3
wrapt==1.15.0
yarl==1.9.2
